First version of support for Oracle datastores in the GUI.
Changed also _getAnnotations(String name) in DatabaseDocumentImpl to read first
the annotations in a transient set, then construct the persistent set from
the transient one, so that the annotations get added to the transient one,
without being put in the list of added annotations. This is the same idea
as the one used in adopt().