Fixed the problem with the Factory initialisation of the documents when adopt-ed in a DS.
The problem was that Marin was trying to initialise the docs through Factory
also when they were read back from the DB, as a result of a getLR call. However
in this case the LRs should be constructed with their constructors, since getLR is
called by Factory.

The only time when documents/corpora need to be created via Factory is in the
adopt() code, because then a new LR is replacing the transient LR. Will be
documented better.