Gate.writeUserConfig respects non-default user config location.
Previously, if the user config file location was overridden, either via the
gate.user.config system property, or by a call to Gate.setUserConfigFile, the
config data would be loaded from this file but saved to the default location
${user.home}/[.]gate.xml.  Now the data is saved to the overridden location.

Gate.getUserConfigFileName is now deprecated.  getDefaultUserConfigFileName has
been introduced, which does the same thing, but has a more accurate name.

This closes bug #1282827.