Changed gate.util.Files.get{Gate}ResourceAsStream to look up resources via the GATE ClassLoader if possible.  This means that resource files like JAPE test cases, etc. can now live in plugin JAR files as well as in gate.jar.
Also added getResource/getGateResource to return a URL to a resource, like
Class.getResource, but again via the GATE ClassLoader if possible.  Icons
are now loaded using this method, so a plugin can specify its own icon
loaded from the plugin's JAR file.  If the <ICON> entry in creole.xml
starts with a '/', it is assumed to be an absolute path.  Otherwise it
is looked up relative to gate/resources/img as currently.

Closes feature request #1505958