Changed SimpleAnnotationSet to extend Set<Annotation> instead of just plain Set.  This allows constructs like:
AnnotationSet as = ...
for(Annotation a : as) {
  ...
}

as well as making the code more self-documenting.

Methods which take or return Sets of annotation types have also been changed to
use Set<String> rather than plain Set.

This change is binary-compatible with the previous version as far as clients of
AnnotationSet are concerned, but alternative implementations of
AnnotationSet/extensions of AnnotationSetImpl will need to be updated.