Added methods for an alternative approach to benchmark logging.  There are now two options:
- startPoint() returns a start time, which you then pass to checkPoint(long,
  String, Object, Map) which logs the event with its duration.
- startPoint(String) logs the start of the event with the given ID, and
  checkPoint(String, Object, Map) logs the end separately.  It is up to the log
  analyser to calculate the duration.

The first form is preferable when the start and end of the operation are close
together in the code and it is easy to hold the start time (e.g. in a local
variable) to pass to checkPoint().  The second form is available for cases
where the start and end are separated and it is difficult to pass the start
time from one to the other.