Changed HTML document format handler to respect the "add space on markup unpack" configuration option, the same as the XML handler already does. Previously, given HTML like:
....first column</td><td>second column...

GATE would always add a space between "column" and "second", even though there
is no space there in the original HTML.  Now, the space will only be added if
the "add space..." configuration option is true.  If it is false, no space will
be added and the document will contain "...first columnsecond column...".