A fairly large set of changes. Highlights include:
We now use a JSplitPane as the main layout of the application editor. This stops the sizes from being updated dynamically, and the various widgets from jumping around the screen when the selected PR is changed.

Better dealing with the strategy panel when editing conditional controllers. In the previous version, sometimes old values remained displayed.

Better sizing of the two tables at the top: they can now take advantage of more space, if available. 

Simpler selection model for member PRs. We are now listening to the selection of the members table, and update all the relevant UI components on those events. The updates used to be distributed over multiple events. This leads to more consistent look for the UI (i.e. if there is no PR selected, then ALL relevant UI elements are emptied).

Editing the strategy pane text fields (feature name and feature value) now automatically switches to "conditional" mode.