Added ControllerAwarePR support to Groovy script PR.  If the script declares a method
void beforeCorpus(c) {
  ...
}

then this will be called before the first document in the corpus is processed.
Similarly afterCorpus(c) will be called after the last document has been
processed, and aborted(c) will be called if the controller fails with an
exception.  In all cases, the "c" parameter is the corpus being processed.  The
callbacks are not executed if the script PR is run standalone or as part of a
simple pipeline (as opposed to a corpus pipeline).