My latest "bored on the train" project - a controller implementation where the execution strategy is specified using a Groovy DSL.
The DSL is very simple.  To execute a particular PR, use the PR's name as a
method call:

myTokeniser()
"ANNIE POS Tagger"()

If there is more than one PR with the same name, they are all run in the order
they appear in the controller's PR list.  To run all the PRs in order,
SerialController style, use allPRs().  You can pass a Map to temporarily
override runtime parameters (the original values are restored after the PR
has run):

mainTransducer(outputASName:"${annotatorName}.out")

To iterate over the documents in the controller's corpus, use eachDocument:

eachDocument {
  theTokeniser()
}

Within the scope of an eachDocument closure, the current document will be
automatically passed to any LanguageAnalyser PRs that are called.

The corpus (if any) over which the controller is running is available as the
variable "corpus", and the current document as "doc".  So you can do
conditional and aggregate things like:

def totalToks = 0
eachDocument {
  // no need to run the tokeniser if we already have tokens
  if(!doc.annotations["Token"]) {
    tokeniser()
  }

  totalToks += doc.annotations["Token"].size()
}
corpus.features.tokens = totalToks

The default script is eachDocument { allPRs() }, which behaves exactly like a
normal SerialAnalyserController.