An attempt at implementing feature value backreferences in JAPE.  This supports syntax like:
Rule: SomeRule
(
  ({Token}):firstTok
  {Token}[0,5]
  {Token.string == :firstTok.Token.string}
):all
-->
:all.Match = {}

Backreferences can be :binding.Type.feature for a feature value,
:binding.Type@metaprop for annotation meta-properties (@string, @length, etc.)
or :binding@string or :binding@length for the covered string or length of a
whole binding AnnotationSet.

Committing on a branch as this will need review and more discussion before we
decide whether to push it to the trunk.  In particular it has important
implications for the alternative JAPE implementations such as jpdaplus.