The FG can now successfully map Lookups from the temp document back to the real one even if they do not exactly line up with the input annotation (e.g., Token) boundaries.  This allows matching to TreeTagger output like Token.lemma == "Spur|Spuren".
Searching is still done with Arrays.binarySearch(...) for speed, but
this is carried out in the dedicated data structure
FlexGazMappingTable (which stores temp/real offsets & calculates
output offsets that must line up with input ones).  A match
transferred from the temp document is extended left/right as necessary
to match input annotations' start/end offsets (respectively) in the
real document.

Restored the suppression of output Lookups that fall entirely outside
the input annotations boundaries (to deal with sections of text that
have been deliberately omitted from the AST, for example).