First cut of a utility class to output GATE document text and annotations in a JSON format (deliberately) close to that used by Twitter to represent entities in tweets:
{
  "text":"Text of the document ....",
  "entities":{
    "Person":[
      {"indices":[0,10], "gender":"male"},
      {"indices":[25,32], "gender":"female"}
    ],
    "Location":[
      {"indices":[75,88], "locType":"country"}
    ]
  }
}

Each annotation is represented as an object with "indices":[start,end] and the
annotation's features represented as other properties of the object.

Also includes support for (more or less) preserving the existing JSON structure
if the original document was a Tweet.

Note that this pulls the Jackson dependency up from the Twitter plugin into
gate-core, so I've removed the conflicting libraries from the plugin.