Added the concept of a "streaming" input handler.
Rather than fetching individual documents from some source given a document ID,
streaming handlers operate like an iterator - each time you call nextDocument()
the next available document will be returned.  Streaming input handlers do not
require a separate document enumerator or pre-defined list of document IDs, but
they require that all documents are loaded in a single thread - you can't
parallelize the loading of different documents.

The motivating use case for this is the format in which data comes from the
Twitter or DataSift streaming APIs, with one file that contains a continuous
stream of JSON objects one after the other.  JSONStreamingInputHandler can load
these kinds of files, treating each JSON object in the stream as a separate
GATE document.  For this to work the JSON format has to be one for which a
DocumentFormat implementation exists to unpack the text from the appropriate
property of the JSON.